Rename catalog core files & minor revisions
