+++ production/db/storage_engine/src/catalog_core.cpp
@@ -2,7 +2,7 @@
 // Copyright (c) Gaia Platform LLC
 // All rights reserved.
 /////////////////////////////////////////////
-#include "light_catalog.hpp"
+#include "catalog_core.hpp"
 
 #include <optional>
