+++ production/db/storage_engine/src/payload_diff.cpp
@@ -6,10 +6,10 @@
 
 #include <mutex>
 
+#include "catalog_core.hpp"
 #include "data_holder.hpp"
 #include "field_access.hpp"
 #include "gaia_common.hpp"
-#include "light_catalog.hpp"
 #include "retail_assert.hpp"
 
 namespace gaia
@@ -24,7 +24,7 @@ private:
     // The map used to store ids of the gaia_table records that define the corresponding types.
     std::unordered_map<gaia_type_t, gaia_id_t> m_type_id_record_id_map;
 
-    void init_type_table_map()
+    void init_type_id_record_id_map()
     {
         for (auto table_view : gaia::db::catalog_core_t::list_tables())
         {
@@ -36,7 +36,7 @@ public:
     // Return the id of the gaia_table record that defines a given type.
     gaia_id_t get_record_id(gaia_type_t type_id)
     {
-        std::call_once(m_type_id_record_id_map_init_flag, &type_id_record_id_cache_t::init_type_table_map, this);
+        std::call_once(m_type_id_record_id_map_init_flag, &type_id_record_id_cache_t::init_type_id_record_id_map, this);
         return m_type_id_record_id_map.at(type_id);
     }
 };
@@ -51,8 +51,8 @@ void compute_payload_diff(
     // Make sure caller passes valid pointer to changed_fields.
     retail_assert(changed_fields);
 
-    static type_id_record_id_cache_t type_table_cache;
-    gaia_id_t type_record_id = type_table_cache.get_record_id(type_id);
+    static type_id_record_id_cache_t type_id_record_id_cache;
+    gaia_id_t type_record_id = type_id_record_id_cache.get_record_id(type_id);
 
     auto schema = catalog_core_t::get_table(type_record_id).binary_schema();
